refactor templates and helper methods and create config file
