+++ src/config.ts
@@ -1,0 +1,3 @@
+export const projectName = process.argv[2];
+export const projectPath = `./${projectName}`;
+export const srcPath = `${projectPath}/src`;
