--- src/helpers/create-src.ts
@@ -1,12 +1,0 @@
-import { createDirectory, createConfigFile } from './index';
-
-/**
- * Create the src file where the main index.ts and index.test.ts files live.
- */
-export function createSrc(dir: string) {
-  const srcDir = `${dir}/src`;
-
-  createDirectory(srcDir);
-  // createConfigFile(dir, `index.ts.txt`, '/src/index.ts');
-  // createConfigFile(dir, `index.test.ts.txt`, '/src/index.test.ts');
-}
