--- src/helpers/index.ts
@@ -1,6 +1,0 @@
-// import { createAllConfigs } from './create-all-configs';
-import { createConfigFile } from './create-config-file';
-import { createDirectory } from './create-directory';
-import { createSrc } from './create-src';
-
-export { createConfigFile, createDirectory, createSrc };
