+++ src/index.ts
@@ -1,23 +1,16 @@
 #!/usr/bin/env node
 import { execSync } from 'child_process';
-import { join } from 'path';
-import { createDirectory, createSrc, createConfigFile } from './helpers';
-import { templates } from './templates';
-import { readFileSync, writeFileSync } from 'fs';
+import { createConfigFile } from './helpers/create-config-file';
+import { createDirectory } from './helpers/create-directory';
+import { projectDirPath } from './helpers/project-dir-path';
+import { srcPath } from './config';
 
 /**
  * Main method for running all the code.
  */
 (function main() {
-  const projectName = process.argv[2];
-
-  if (!projectName) {
-    console.error('Please provide a name for the project');
-    return;
-  }
-
-  const newDirectory = `${process.cwd()}/${projectName}`;
-  createDirectory(newDirectory);
-  createConfigFile(newDirectory);
-  // execSync(`cd ${projectName} && npm install`, { stdio: [0, 1, 2] });
+  createDirectory(projectDirPath());
+  createDirectory(srcPath);
+  createConfigFile();
+  execSync(`cd ${projectDirPath()} && npm install`, { stdio: [0, 1, 2] });
 })();
